Move quadratic_knapsack to knapsack.py
