+++ docs/reference/generators.rst
@@ -55,7 +55,6 @@ Optimization
    random_bin_packing
    random_knapsack
    random_multi_knapsack
-   random_quadratic_knapsack
 
 Random
 ======
@@ -72,7 +71,6 @@ Random
    random_knapsack
    random_multi_knapsack
    random_nae3sat
-   random_quadratic_knapsack
    ran_r
    uniform
