+++ releasenotes/notes/quadratic-knapsack-generator-55cb808252579b38.yaml
@@ -2,4 +2,3 @@
 features:
   - |
     Add ``quadratic_knapsack`` generator function to ``dimod.generators``.
-    Add ``random_quadratic_knapsack`` generator function to ``dimod.generators``.
